requirements.txt
Flask>=2.2.0,<3.0.0
Flask-SQLAlchemy>=3.0.2,<4.0.0
psycopg2-binary>=2.8.6,<3.0.0  
Flask-Migrate>=3.1.0,<4.0.0
Flask-Script>=2.0.0,<3.0.0

Flask-WTF>=0.15.0,<1.0.0
Flask-Bootstrap>=3.3.0,<4.0.0
Flask-Login>=0.6.0,<1.0.0
Flask-Mail>=0.9.0,<1.0.0
Flask-Bcrypt>=0.7.0,<1.0.0
Flask-Caching>=1.10.0,<2.0.0
python-dotenv>=0.10.0,<0.11.0
gunicorn>=20.0,<21.0

Flask-JWT-Extended>=4.0.0,<5.0.0
flask-restful-swagger>=0.2.0,<0.3.0
